refactor: 공통로직분리( 파일다운 controller, Mapper 인터페이스)
