+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>UnknownProject2</name>
+	<name>UnknownProject6</name>
 	<comment></comment>
 	<projects>
 	</projects>
