+++ src/main/java/com/unknown/paldak/admin/mapper/BaseMapper.java
@@ -1,0 +1,24 @@
+package com.unknown.paldak.admin.mapper;
+
+
+import java.util.List;
+
+import com.unknown.paldak.admin.common.domain.Criteria;
+
+public interface BaseMapper<T> {
+    List<T> getList();
+
+    void insert(T item);
+
+    void insertSelectKey(T item);
+
+    T read(Long id);
+
+    int delete(Long id);
+
+    int update(T item);
+
+    List<T> getListWithPaging(Criteria cri);
+
+    int getTotalCount(Criteria cri);
+}
