+++ src/main/webapp/WEB-INF/views/admin/includes/header.jsp
@@ -38,3 +38,4 @@
 </head>
 
 <body>
+ <a href="/">홈</a>
