+++ src/main/webapp/WEB-INF/views/admin/product/modify.jsp
@@ -48,7 +48,7 @@
 						</div>
 						<div class="form-group">
 							<label>변경전 이미지</label>
-							<img src="upload/${product.productImageURL}" alt="제품이미지" style="max-width: 100px">
+							<img src="/download/${product.productImageURL}" alt="제품이미지" style="max-width: 100px">
 							<input type="hidden" value='${product.productImageURL}' name='productImageURL'>
 							
 						</div>
