+++ src/main/webapp/WEB-INF/views/home.jsp
@@ -17,5 +17,7 @@
 <a href="admin/qna/list">관리자 qna</a>
 <a href="admin/faq/list">관리자 faq</a>
 <a href="admin/review/list">관리자 review</a>
+<a href="productDetails">테스트</a>
+
 </body>
 </html>
